services/common/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1
boto3==1.34.19
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==1.10.14
email-validator==2.2.0
python-multipart